+++ components/molecules/SuperlativeSelector/superlative-selector.tsx
@@ -1,18 +1,24 @@
 import ContextFilterButton from "components/atoms/ContextFilterButton/context-filter-button";
-import ContextFilterOption from "components/atoms/ContextFilterOption/context-filter-option";
+
 import React, { useEffect, useRef, useState } from "react";
 import Icon from "../../atoms/Icon/icon";
 import cancelIcon from "public/x-circle.svg";
 import Radio from "components/atoms/Radio/radio";
+import humanizeNumber from "lib/utils/humanizeNumber";
+import getFilterKey from "lib/utils/get-filter-key";
+
 interface SuperlativeSelectorProps {
   filterOptions: string[];
+  filterValues: { [name: string]: number | undefined };
   handleFilterClick: (filter: string) => void;
   handleCancelClick: () => void;
   className?: string;
   selected?: string;
 }
+
 const SuperativeSelector: React.FC<SuperlativeSelectorProps> = ({
   filterOptions,
+  filterValues,
   handleFilterClick,
   handleCancelClick,
   selected
@@ -42,13 +48,13 @@ const SuperativeSelector: React.FC<SuperlativeSelectorProps> = ({
 
   return (
     <div className="max-w-max relative" ref={ref}>
-      <ContextFilterButton isSelected={selected ? true : false}>
+      <ContextFilterButton onClick={toggleFilter} isSelected={!!selected}>
         {selected ? (
           <div className="flex">
             <div className="flex" onClick={toggleFilter}>
               <span className="text-dark-slate-10">Filtered by:</span>
               <div className="ml-1 text-light-slate-12">
-                {filterOptions.find((option) => option.toLowerCase().replaceAll(" ", "-") === selected)}
+                {filterOptions.find((option) => getFilterKey(option) === selected)}
               </div>
             </div>
             <Icon
@@ -56,31 +62,36 @@ const SuperativeSelector: React.FC<SuperlativeSelectorProps> = ({
               IconImage={cancelIcon}
               onClick={() => {
                 handleCancelClick();
-                toggleFilter();
+                setIsOpen(false);
               }}
             />
           </div>
         ) : (
-          <div onClick={toggleFilter}>Add Filter</div>
+          <div onClick={toggleFilter}>Add Filter </div>
         )}
       </ContextFilterButton>
       {isOpen && (
-        <div className="absolute -left-full md:left-0 space-y-1 mt-1 shadow-superlative w-64 z-10 bg-white rounded-lg px-1.5 py-2">
+        <div className="absolute -left-full md:left-0 space-y-1 mt-1 shadow-superlative w-72 z-10 bg-white rounded-lg px-1.5 py-2">
           {filterOptions.length > 0 &&
-            filterOptions.map((option, index) => (
-              <Radio
-                withLabel={true}
-                key={index}
-                onClick={() => {
-                  handleFilterClick(option.replaceAll(" ", "-"));
-                  toggleFilter();
-                }}
-                css="!w-full"
-                checked={selected === option.toLocaleLowerCase().replaceAll(" ", "-") ? true : false}
-              >
-                {option}
-              </Radio>
-            ))}
+            filterOptions.map((option, index) => {
+              const filterKey = getFilterKey(option);
+              const filterValue = filterValues[filterKey];
+
+              return (
+                <Radio
+                  withLabel={filterValue ? humanizeNumber(filterValue, "abbreviation") : "-"}
+                  key={index}
+                  onClick={() => {
+                    handleFilterClick(filterKey);
+                    toggleFilter();
+                  }}
+                  css="!w-full"
+                  checked={selected === filterKey ? true : false}
+                >
+                  {option}
+                </Radio>
+              );}
+            )}
         </div>
       )}
     </div>