+++ CHANGELOG.md
@@ -5,6 +5,27 @@
 
 > All notable changes to this project will be documented in this file
 
+## [1.10.0-beta.3](https://github.com/open-sauced/insights/compare/v1.10.0-beta.2...v1.10.0-beta.3) (2022-10-04)
+
+
+### 🐛 Bug Fixes
+
+* Click target on filter button icon added ([#483](https://github.com/open-sauced/insights/issues/483)) ([d4aba03](https://github.com/open-sauced/insights/commit/d4aba0379152f394deeb2f26e0bf99679efc7a8c))
+
+## [1.10.0-beta.2](https://github.com/open-sauced/insights/compare/v1.10.0-beta.1...v1.10.0-beta.2) (2022-10-04)
+
+
+### 🍕 Features
+
+* connect dashboard insights to the API ([#485](https://github.com/open-sauced/insights/issues/485)) ([f4e5b6e](https://github.com/open-sauced/insights/commit/f4e5b6e854cf1422a3b3d17272dd659ccbbcaf1c)), closes [#481](https://github.com/open-sauced/insights/issues/481)
+
+## [1.10.0-beta.1](https://github.com/open-sauced/insights/compare/v1.9.0...v1.10.0-beta.1) (2022-10-04)
+
+
+### 🍕 Features
+
+* prefetch totals for filter options ([#479](https://github.com/open-sauced/insights/issues/479)) ([a97b09b](https://github.com/open-sauced/insights/commit/a97b09b7f650d79114b0e8fec841087fce73e278)), closes [#471](https://github.com/open-sauced/insights/issues/471)
+
 ## [1.9.0](https://github.com/open-sauced/insights/compare/v1.8.0...v1.9.0) (2022-10-03)